SET-559 refactor and minor optimization to minimize code duplications
